Change type for the Sequence field to int64 (#81)

+++ product.go
@@ -62,7 +62,7 @@ type BookEntry struct {
 }
 
 type Book struct {
-	Sequence int         `json:"sequence"`
+	Sequence int64       `json:"sequence"`
 	Bids     []BookEntry `json:"bids"`
 	Asks     []BookEntry `json:"asks"`
 }
